Reorganized and refactored the code.
